# build/swagger.json
{
	"components": {
...
		"schemas": {
...
			"IngredientDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"enName": {
						"type": "string"
					},
					"proof": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"enName",
					"proof",
					"category"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IngredientDetailDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"enName": {
						"type": "string"
					},
					"proof": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					},
					"quantity": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"enName",
					"proof",
					"category",
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "nodeProjectTemplate",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/search": {
			"get": {
				"operationId": "GetSearchKeywords",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"messsage": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"messsage",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"ingredients": {
													"items": {
														"$ref": "#/components/schemas/IngredientDTO"
													},
													"type": "array"
												},
												"recipes": {
													"items": {
														"properties": {
															"defaultRecipe": {
																"type": "string"
															},
															"ingredients": {
																"items": {
																	"$ref": "#/components/schemas/IngredientDetailDTO"
																},
																"type": "array"
															},
															"steps": {
																"items": {
																	"type": "string"
																},
																"type": "array"
															},
															"glass": {
																"properties": {
																	"url": {
																		"type": "string"
																	},
																	"name": {
																		"type": "string"
																	}
																},
																"required": [
																	"url",
																	"name"
																],
																"type": "object"
															},
															"skill": {
																"properties": {
																	"url": {
																		"type": "string"
																	},
																	"name": {
																		"type": "string"
																	}
																},
																"required": [
																	"url",
																	"name"
																],
																"type": "object"
															},
															"proofIcon": {
																"type": "string"
															},
															"proof": {
																"type": "number",
																"format": "double"
															},
															"base": {
																"properties": {
																	"url": {
																		"type": "string"
																	},
																	"name": {
																		"type": "string"
																	},
																	"id": {
																		"type": "string"
																	}
																},
																"required": [
																	"url",
																	"name",
																	"id"
																],
																"type": "object"
															},
															"enName": {
																"type": "string"
															},
															"name": {
																"type": "string"
															},
															"id": {
																"type": "string"
															}
														},
														"required": [
															"defaultRecipe",
															"ingredients",
															"steps",
															"glass",
															"skill",
															"proofIcon",
															"proof",
															"base",
															"enName",
															"name",
															"id"
														],
														"type": "object"
													},
													"type": "array"
												},
												"recommends": {
													"items": {
														"properties": {
															"name": {},
															"recipeId": {}
														},
														"required": [
															"name",
															"recipeId"
														],
														"type": "object"
													},
													"type": "array"
												}
											},
											"required": [
												"ingredients",
												"recipes",
												"recommends"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/search/base": {
			"get": {
				"operationId": "FindRecipesByBase",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"recipes": {
											"items": {
												"properties": {
													"defaultRecipe": {
														"type": "string"
													},
													"ingredients": {
														"items": {
															"$ref": "#/components/schemas/IngredientDetailDTO"
														},
														"type": "array"
													},
													"steps": {
														"items": {
															"type": "string"
														},
														"type": "array"
													},
													"glass": {
														"properties": {
															"url": {
																"type": "string"
															},
															"name": {
																"type": "string"
															}
														},
														"required": [
															"url",
															"name"
														],
														"type": "object"
													},
													"skill": {
														"properties": {
															"url": {
																"type": "string"
															},
															"name": {
																"type": "string"
															}
														},
														"required": [
															"url",
															"name"
														],
														"type": "object"
													},
													"proofIcon": {
														"type": "string"
													},
													"proof": {
														"type": "number",
														"format": "double"
													},
													"base": {
														"properties": {
															"url": {
																"type": "string"
															},
															"name": {
																"type": "string"
															},
															"id": {
																"type": "string"
															}
														},
														"required": [
															"url",
															"name",
															"id"
														],
														"type": "object"
													},
													"enName": {
														"type": "string"
													},
													"name": {
														"type": "string"
													},
													"id": {
														"type": "string"
													}
												},
												"required": [
													"defaultRecipe",
													"ingredients",
													"steps",
													"glass",
													"skill",
													"proofIcon",
													"proof",
													"base",
													"enName",
													"name",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										}
									},
									"required": [
										"recipes"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "base",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/search/keyword": {
			"get": {
				"operationId": "SearchRecipesByKeyword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"recipes": {
											"items": {
												"properties": {
													"defaultRecipe": {
														"type": "string"
													},
													"ingredients": {
														"items": {
															"$ref": "#/components/schemas/IngredientDetailDTO"
														},
														"type": "array"
													},
													"steps": {
														"items": {
															"type": "string"
														},
														"type": "array"
													},
													"glass": {
														"properties": {
															"url": {
																"type": "string"
															},
															"name": {
																"type": "string"
															}
														},
														"required": [
															"url",
															"name"
														],
														"type": "object"
													},
													"skill": {
														"properties": {
															"url": {
																"type": "string"
															},
															"name": {
																"type": "string"
															}
														},
														"required": [
															"url",
															"name"
														],
														"type": "object"
													},
													"proofIcon": {
														"type": "string"
													},
													"proof": {
														"type": "number",
														"format": "double"
													},
													"base": {
														"properties": {
															"url": {
																"type": "string"
															},
															"name": {
																"type": "string"
															},
															"id": {
																"type": "string"
															}
														},
														"required": [
															"url",
															"name",
															"id"
														],
														"type": "object"
													},
													"enName": {
														"type": "string"
													},
													"name": {
														"type": "string"
													},
													"id": {
														"type": "string"
													}
												},
												"required": [
													"defaultRecipe",
													"ingredients",
													"steps",
													"glass",
													"skill",
													"proofIcon",
													"proof",
													"base",
													"enName",
													"name",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										}
									},
									"required": [
										"recipes"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "keyword",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}